Improved layout for landscape mode in size tab removed decimals from display of sizes

+++ gic_flutter/lib/src/views/screenEditor/controlDialog/textTab.dart
@@ -112,7 +112,7 @@ class TextTabState extends BaseTabState {
   TextField _size() {
     TextEditingController controller = TextEditingController();
     textControllers.add(controller);
-    controller.text = widget.gicEditControl.control.font.size.toString();
+    controller.text = widget.gicEditControl.control.font.size.toInt().toString();
     controller.addListener(() {
       widget.gicEditControl.control.font.size = double.parse(controller.text);
     });
